requirements.txt
fastapi
uvicorn[standard]
pydantic
WebSocket
fastapi.responses
python-dotenv
google-generativeai